requirements.txt
streamlit==1.41.1
streamlit-code-editor==0.1.21